use logging datastore for sql in validate
