+++ internal/pkg/core/kelon.go
@@ -107,87 +107,6 @@ func (k *Kelon) Start() {
 	k.stopOnSIGTERM()
 }
 
-func (k *Kelon) StartValidate() {
-	if !k.configured {
-		k.logger.Fatalf("Kelon was not configured! Please call Configure()!")
-	}
-
-	k.makeConfigWatcher(configs.FileConfigLoader{}, k.config.ConfigWatcherPath)
-
-	confLoader := configs.FileConfigLoader{FilePath: *k.config.ConfigPath}
-	loadedConf, err := confLoader.Load()
-	if err != nil {
-		k.logger.Fatalln("Unable to parse configuration: ", err.Error())
-	}
-
-	// Create logging file
-	if k.config.QueryOutputFilename != nil && *k.config.QueryOutputFilename != "" {
-		f, fileErr := os.Create(*k.config.QueryOutputFilename)
-		if fileErr != nil {
-			k.logger.Fatalln("Unable to parse crate file: ", fileErr.Error())
-		}
-
-		defer f.Close()
-		k.dsLoggingWriter = f
-	}
-
-	ctx := context.Background()
-	// Configure application
-	var (
-		config     = new(configs.AppConfig)
-		compiler   = opaInt.NewPolicyCompiler()
-		parser     = requestInt.NewURLProcessor()
-		mapper     = requestInt.NewPathMapper()
-		translator = translateInt.NewAstTranslator()
-	)
-
-	// Build config
-	config.APIMappings = loadedConf.APIMappings
-	config.DatastoreSchemas = loadedConf.DatastoreSchemas
-	config.Datastores = loadedConf.Datastores
-	config.OPA = loadedConf.OPA
-	config.MetricsProvider = telemetry.NewNoopMetricProvider()
-	k.metricsProvider = config.MetricsProvider // Stopped gracefully later on
-	config.TraceProvider = telemetry.NewNoopTraceProvider()
-	k.traceProvider = config.TraceProvider // Stopped gracefully later on
-
-	serverConf := k.makeServerConfig(compiler, parser, mapper, translator, loadedConf)
-
-	config.CallOperands, err = data.LoadAllCallOperands(config.Datastores, k.config.OperandDir)
-	if err != nil {
-		k.logger.Fatalln(err.Error())
-	}
-
-	err = (*serverConf.Compiler).Configure(config, &serverConf.PolicyCompilerConfig)
-	if err != nil {
-		k.logger.Fatalln(err.Error())
-	}
-
-	body, err := parseBodyFromString(*k.config.InputBody)
-	if err != nil {
-		k.logger.Fatalln(err.Error())
-	}
-
-	// Execute Request
-	d, err := compiler.Execute(ctx, body)
-	if err != nil {
-		k.logger.Fatalln(err.Error())
-	}
-
-	var allowString string
-	if d.Allow {
-		allowString = "ALLOW"
-	} else {
-		allowString = "DENY"
-	}
-
-	logFields := log.Fields{
-		logging.LabelPath:   d.Path,
-		logging.LabelMethod: d.Method,
-	}
-	logging.LogAccessDecision(serverConf.AccessDecisionLogLevel, allowString, "main", logFields)
-}
-
 func (k *Kelon) onConfigLoaded(change watcher.ChangeType, loadedConf *configs.ExternalConfig, err error) {
 	if err != nil {
 		k.logger.Fatalln("Unable to parse configuration: ", err.Error())
